2.5.0 Aide CmdLog

+++ commands/Admin/aideadm.js
@@ -78,7 +78,7 @@ class AideAdmCommand extends Command {
                 true,
             );
 
-        message.delete();
+        if (message.channel.type === 'text') if (message.channel.type === 'text') message.delete();;;
         return message.util.send(embed);
 
     }
