+++ commands/Modération/logs.js
@@ -6,12 +6,24 @@ const { Permissions } = require('discord.js');
 class LogsCommand extends Command {
     constructor() {
         super('logs', {
-            aliases: ['logs'],
-            userPermissions: [Permissions.FLAGS.MANAGE_MESSAGES],
+            aliases: ['logs', 'log'],  
             category: 'Modération',
+            channelRestriction: 'dm',
             cooldown: 30000,
             ratelimit: 1,
-            description: 'Récupérer les logs',
+            description: {
+                content: 'Consultation des logs',
+                usage: '!logs [action]\nActions: \`full\`, \`commandes\` ( ou \`cmds\`)',
+                examples: ['!logs', '!logs full', '!logs cmds']
+            },
+            args: [{
+                id: "action",
+                type: [
+                    "full",
+                    "commandes", "cmds"
+                ],
+                default: "full",
+            }],
         });
     }
 
@@ -20,24 +32,13 @@ class LogsCommand extends Command {
         const guild = client.guilds.get(client.config.guildID);
         const settings = await client.db.getSettings(client);
 
-        message.delete();
-        message.channel.send("Logs", {files: ["/root/.pm2/logs/index-out.log"]});
-        //message.channel.send("Logs", {files: ["C:\\Users\\Bad\\Documents\\AlfredProd\\data\\enmap.sqlite"]});
+        if (message.channel.type === 'text') message.delete();;
+        message.channel.send("Logs", { files: ["/root/.pm2/logs/index-out.log"] });
 
-
-        /*
-        switch (args.action) {
-            case 'dl':
-
-            break;
-            case 'debug':
-
-            break;
-        }
-        */
-
+        if (message.channel.type === 'text') if (message.channel.type === 'text') message.delete();
 
     }
+
 }
