+++ commands/Utilitaires/dé.js
@@ -2,10 +2,10 @@ const {
     Command
 } = require('discord-akairo');
 
-class TirageCommand extends Command {
+class DéCommand extends Command {
     constructor() {
-        super('tirage', {
-            aliases: ['tirage','random', 'rand', 'rnd', 'dés'],
+        super('dé', {
+            aliases: ['dé','random', 'rand', 'rnd'],
             category: 'Utilitaires',
             // define arg properties
             args: [{
@@ -13,18 +13,27 @@ class TirageCommand extends Command {
                 type: 'number',
                 prompt: {
                     start: 'Combien de faces possède votre dé virtuel ?',
-                    retry: 'Hmmm, je m\'attendais à un nombre...'
+                    retry: 'Hmmm, je m\'attendais à un nombre...',
+                    default: 6
                 }
             }, {
                 id: 'numTirages',
                 type: 'number',
                 prompt: {
                     start: 'Combien de tirages souhaitez vous faire ?',
-                    retry: 'Hmmm, je m\'attendais à un nombre...'
+                    retry: 'Hmmm, je m\'attendais à un nombre...',
+                    default: 1
                 }
             }, ],
             // command description
-            description: 'Lance un dé virtuel.',
+
+            description: {
+                content: 'Permet d\'effectuer un lancer de dé virtuel',
+                usage: '\`!dé [<nb faces> <nb tirages>]\`\n' +
+                'Choisissez le nombre de faces que comporte votre dé et le nombre de tirages successifs que vous souhaitez que j\'effectue.\n' + 
+                'Si vous lancez la commande sans paramètre, j\'effectuerais un seul lancé d\'un dé à six faces.',
+                examples: ['!dé','!dé 20 4', '!rnd 100 1']
+            }
         });
     }
 
@@ -38,10 +47,10 @@ class TirageCommand extends Command {
         }
         msg += (Math.floor(Math.random()*args.numFaces)+1).toString();
         message.reply(msg);
-        message.delete();
+        if (message.channel.type === 'text') message.delete();;
 
     }
 }
 
 
-module.exports = TirageCommand;
+module.exports = DéCommand;
