+++ commands/Utilitaires/suggestion.js
@@ -6,14 +6,26 @@ const Discord = require("discord.js");
 class SuggestionCommand extends Command {
     constructor() {
         super('suggestion', {
-            aliases: ['suggestion', 'sug'],
+            aliases: ['suggestion', 'sugg', 'sug'],
             category: 'Utilitaires',
             args: [
                 {
                     id: 'suggestion',
                     match: 'content'
                 }
-            ]
+            ],
+            description: {
+                content: 'Permet de soumettre une suggestion dans <#688062451228475428>',
+                usage: '\`!suggestion <ma suggestion>\`\n\n' +
+                    'Vous pouvez utiliser cette commande directement dans <#688062451228475428> ou bien directement ici par message privé.\n' +
+                    'Votre suggestion apparaitra dans le salon sous un format particulier avec les réactions 💚💛💔. Les réactions permettront de recceuillir l\'avis des membres, elle auront évidement un poid dans la balance de la décision finale.'
+                    ,
+                examples: [
+                    '!suggestion Que pensez-vous d\'ajouter le mod **Super mod** sur Ark ?',
+                    '!sugg Et si on créait des salons pour **New Game** qui arrive bientôt ?',
+                    '!sug Je propose de créer #Imprimantes-3D pour soulager #media-flood!'
+                ]
+            }
         });
     }
 
@@ -30,7 +42,7 @@ class SuggestionCommand extends Command {
 
         /*
         if (message.channel.type !== "dm") {
-            message.delete();
+            if (message.channel.type === 'text') message.delete();;
         }
         */
 
@@ -51,11 +63,12 @@ class SuggestionCommand extends Command {
         embed.setFooter(client.textes.get("SUGG_NOTIF_PROPOSED_BY", member), avatar);
         embed.setTimestamp();
 
-        message.delete();
-        
+        if (message.channel.type === 'text') message.delete();;
+
         return suggChannel.send(embed).then(async msgSent => {
-            await msgSent.react("✅");
-            await msgSent.react("❌");
+            await msgSent.react("💚");
+            await msgSent.react("💛");
+            await msgSent.react("💔");
         })
 
     }
